Rename helper to `value-clone.ts`.
